ci: fix publish-dockerhub (DAEMONS)
Commit 82b9ba16, part of #816, broke upload of all DAEMONS but
base. The reason why it wasn't detected is because composite action
leaves DAEMONS set, so step 3 passes and step 4 just uploads
the single daemon base.

Fix it by defining it globally, and add a check for the list of
DAEMONS as part of step2, so that this is not happening again.